Exercise 4.21: bloglist expansion, step9

+++ part4/bloglist/controllers/blogs.js
@@ -43,6 +43,21 @@ blogsRouter.get("/:id", async (request, response) => {
 });
 
 blogsRouter.delete("/:id", async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken) {
+    return response.status(401).json({
+      error: "token is missing or invalid",
+    })
+  }
+
+  const blog = await Blog.findById(request.params.id);
+
+  if (decodedToken.id.toString() !== blog.user.toString()) {
+    return response.status(401).json({
+      error: "permission denied"
+    })
+  }
+
   await Blog.findByIdAndDelete(request.params.id);
   response.status(204).end();
 });
